+++ src/output/variables/output_variables.jl
@@ -5,7 +5,9 @@ include("boundaries.jl")      # BoundaryOutput()
 include("radiation.jl")       # RadiationOutput()
 include("stochastic.jl")      # RandomPatternOutput()
 include("surface_fluxes.jl")  # SurfaceFluxesOutput()
+include("land.jl")            # LandOutput()
 include("ocean.jl")           # SeaSurfaceTemperatureOutput()
+include("tracers.jl")         # TracerOutput()
 
 # collect all together for conveneince
 AllOutputVariables() = (
@@ -14,5 +16,6 @@ AllOutputVariables() = (
     RadiationOutput()...,
     RandomPatternOutput(),
     SurfaceFluxesOutput()...,
+    LandOutput()...,
     SeaSurfaceTemperatureOutput(),
 )